chore: clean up types in the GroupChannel and GroupChannelList modules
