+++ src/modules/GroupChannel/index.tsx
@@ -1,15 +1,13 @@
 import React from 'react';
 
-import type { GroupChannelContextProps } from './context/GroupChannelProvider';
-import { GroupChannelProvider } from './context/GroupChannelProvider';
+import { GroupChannelProvider, GroupChannelProviderProps } from './context/GroupChannelProvider';
 import GroupChannelUI, { GroupChannelUIProps } from './components/GroupChannelUI';
 
-export interface GroupChannelProps extends GroupChannelContextProps, GroupChannelUIProps { }
-
+export interface GroupChannelProps extends GroupChannelProviderProps, GroupChannelUIProps {}
 export const GroupChannel = (props: GroupChannelProps) => {
   return (
     <GroupChannelProvider {...props}>
-      <GroupChannelUI {...props}/>
+      <GroupChannelUI {...props} />
     </GroupChannelProvider>
   );
 };
