+++ src/modules/GroupChannelList/index.tsx
@@ -1,16 +1,12 @@
 import React from 'react';
-import {
-  GroupChannelListProvider,
-  GroupChannelListProviderProps,
-} from './context/GroupChannelListProvider';
 
+import { GroupChannelListProvider, GroupChannelListProviderProps } from './context/GroupChannelListProvider';
 import GroupChannelListUI, { GroupChannelListUIProps } from './components/GroupChannelListUI';
 
 export interface GroupChannelListProps extends GroupChannelListProviderProps, GroupChannelListUIProps {}
-
 export const GroupChannelList = (props: GroupChannelListProps) => {
   return (
-    <GroupChannelListProvider {...props} >
+    <GroupChannelListProvider {...props}>
       <GroupChannelListUI {...props} />
     </GroupChannelListProvider>
   );
