+++ src/utils/typeHelpers/partialRequired.ts
@@ -1,0 +1,1 @@
+export type PartialRequired<T, RequiredKeys extends keyof T> = Partial<Omit<T, RequiredKeys>> & Required<Pick<T, RequiredKeys>>;
